Set up infrastructure to write output with a frequency other than steps.
This commit adds a timestamp to the IOControl structure which will
keep a record of the timestamp of the last time output was written.

This can then be used with the current timestamp to determine the number
of secs, minutes, hours, days, months and years that have passed since the
last output was written.